Setup rootlayout and home to work with refactoring
